feat: ComponentScanner 가져오기, 테스트
