+++ README.md
@@ -3,3 +3,7 @@
 ## 할 일들
 
 ### 🚀 0단계 - 기본 코드 준비
+
+- [x] 요구사항 1 - @Entity 엔터티 어노테이션이 있는 클래스만 가져오기
+- [ ] 요구사항 2 - scanner 로 찾은 Entity Class 정보를 통해 MetamodelImpl 에 데이터를 채워넣어보자
+- [ ] 요구사항 3 - Metamodel 을 활용해 리팩터링을 진행해보자
