+++ src/test/java/database/sql/ddl/CreateTest.java
@@ -3,14 +3,13 @@ package database.sql.ddl;
 import database.dialect.Dialect;
 import database.dialect.MySQLDialect;
 import database.mapping.AllEntities;
+import entity.Departure;
+import entity.Employee;
 import entity.Person;
-import jakarta.persistence.*;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -24,9 +23,9 @@ class CreateTest {
 
     @ParameterizedTest
     @CsvSource(value = {
-            "database.sql.ddl.OldPerson1:CREATE TABLE OldPerson1 (id BIGINT PRIMARY KEY, name VARCHAR(255) NULL, age INT NULL)",
-            "database.sql.ddl.OldPerson2:CREATE TABLE OldPerson2 (id BIGINT AUTO_INCREMENT PRIMARY KEY, nick_name VARCHAR(255) NULL, old INT NULL, email VARCHAR(255) NOT NULL)",
-            "database.sql.ddl.OldPerson3:CREATE TABLE users (id BIGINT AUTO_INCREMENT PRIMARY KEY, nick_name VARCHAR(255) NULL, old INT NULL, email VARCHAR(255) NOT NULL)"
+            "entity.OldPerson1:CREATE TABLE OldPerson1 (id BIGINT PRIMARY KEY, name VARCHAR(255) NULL, age INT NULL)",
+            "entity.OldPerson2:CREATE TABLE OldPerson2 (id BIGINT AUTO_INCREMENT PRIMARY KEY, nick_name VARCHAR(255) NULL, old INT NULL, email VARCHAR(255) NOT NULL)",
+            "entity.OldPerson3:CREATE TABLE users (id BIGINT AUTO_INCREMENT PRIMARY KEY, nick_name VARCHAR(255) NULL, old INT NULL, email VARCHAR(255) NOT NULL)"
     }, delimiter = ':')
     void buildCreateQuery(Class<?> clazz, String expected) {
         assertCreateQuery(clazz, expected);
@@ -42,26 +41,6 @@ class CreateTest {
         Create create = new Create(clazz, dialect);
         String actual = create.buildQuery();
         assertThat(actual).isEqualTo(expected);
-    }
-
-    @Entity
-    static class Departure {
-        @Id
-        @GeneratedValue
-        Long id;
-
-        @OneToMany
-        @JoinColumn(name = "departure_id")
-        List<Employee> employees;
-    }
-
-    @Entity
-    static class Employee {
-        @Id
-        @GeneratedValue
-        Long id;
-
-        String name;
     }
 
     @Test
