--- src/test/java/database/sql/dml/NoAutoIncrementUser.java
@@ -1,45 +1,0 @@
-package database.sql.dml;
-
-import jakarta.persistence.*;
-
-@Table(name = "users_no_auto_increment")
-@Entity
-public class NoAutoIncrementUser {
-
-    @Id
-    private Long id;
-
-    @Column(name = "nick_name")
-    private String name;
-
-    @Column(name = "old")
-    private Integer age;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Transient
-    private Integer index;
-
-    public NoAutoIncrementUser() {
-    }
-
-    public NoAutoIncrementUser(Long id, String name, Integer age, String email) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-        this.email = email;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
