+++ src/test/java/database/sql/dml/SelectTest.java
@@ -3,6 +3,7 @@ package database.sql.dml;
 import database.mapping.EntityMetadata;
 import database.mapping.EntityMetadataFactory;
 import database.sql.dml.part.WhereMap;
+import entity.Person4;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
