+++ src/test/java/entity/EagerLoadTestOrderItem.java
@@ -1,10 +1,10 @@
-package persistence.entity.testentities;
+package entity;
 
 import jakarta.persistence.*;
 
 @Entity
-@Table(name = "lazyload_order_items")
-public class LazyLoadTestOrderItem {
+@Table(name = "eagerload_order_items")
+public class EagerLoadTestOrderItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -14,22 +14,18 @@ public class LazyLoadTestOrderItem {
 
     private Integer quantity;
 
-//    // 테이블 생성때문에 필요함
-//    @Column(name = "order_id", nullable = false)
-//    private Long orderId;
-
-    public LazyLoadTestOrderItem(String product, int quantity) {
+    public EagerLoadTestOrderItem(String product, int quantity) {
         this.product = product;
         this.quantity = quantity;
     }
 
-    public LazyLoadTestOrderItem(String product, Integer quantity, Long orderId) {
+    public EagerLoadTestOrderItem(String product, Integer quantity, Long orderId) {
         this.product = product;
         this.quantity = quantity;
 //        this.orderId = orderId;
     }
 
-    public LazyLoadTestOrderItem() {
+    public EagerLoadTestOrderItem() {
     }
 
     public Long getId() {
