+++ src/test/java/entity/Employee.java
@@ -1,0 +1,14 @@
+package entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+
+@Entity
+public class Employee {
+    @Id
+    @GeneratedValue
+    Long id;
+
+    String name;
+}
