+++ src/test/java/entity/NoAutoIncrementUser.java
@@ -1,13 +1,12 @@
-package database.sql.dml;
+package entity;
 
 import jakarta.persistence.*;
 
-@Table(name = "users")
+@Table(name = "users_no_auto_increment")
 @Entity
-public class Person4 {
+public class NoAutoIncrementUser {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(name = "nick_name")
@@ -21,6 +20,16 @@ public class Person4 {
 
     @Transient
     private Integer index;
+
+    public NoAutoIncrementUser() {
+    }
+
+    public NoAutoIncrementUser(Long id, String name, Integer age, String email) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.email = email;
+    }
 
     public void setName(String name) {
         this.name = name;
