+++ src/test/java/entity/OldPerson1.java
@@ -1,4 +1,4 @@
-package database.sql.ddl;
+package entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
