+++ src/test/java/entity/OldPerson2.java
@@ -1,0 +1,21 @@
+package entity;
+
+import jakarta.persistence.*;
+
+@Entity
+public class OldPerson2 {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "nick_name")
+    private String name;
+
+    @Column(name = "old")
+    private Integer age;
+
+    @Column(nullable = false)
+    private String email;
+
+}
