+++ src/test/java/persistence/bootstrap/ComponentScannerTest.java
@@ -1,0 +1,29 @@
+package persistence.bootstrap;
+
+import entity.*;
+import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class ComponentScannerTest {
+    ComponentScanner componentScanner = new ComponentScanner();
+
+    @Test
+    void scan() throws IOException, ClassNotFoundException {
+        List<Class<?>> result = componentScanner.scan("entity");
+
+        assertThat(result).containsAll(
+                List.of(
+                        Departure.class,
+                        Employee.class,
+                        OldPerson1.class,
+                        OldPerson2.class,
+                        OldPerson3.class,
+                        Person4.class
+                ));
+    }
+
+}
